+++ src/PimEnterprise/Bundle/DataGridBundle/DependencyInjection/PimEnterpriseDataGridExtension.php
@@ -29,6 +29,7 @@ class PimEnterpriseDataGridExtension extends Extension
     public function load(array $configs, ContainerBuilder $container)
     {
         $loader = new YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
+        $loader->load('actions.yml');
         $loader->load('configurators.yml');
         $loader->load('controllers.yml');
         $loader->load('data_sources.yml');