+++ src/PimEnterprise/Bundle/WorkflowBundle/Doctrine/ORM/Repository/PublishedProductRepository.php
@@ -13,6 +13,7 @@ namespace PimEnterprise\Bundle\WorkflowBundle\Doctrine\ORM\Repository;
 
 use Doctrine\ORM\NoResultException;
 use Doctrine\ORM\QueryBuilder;
+use Pim\Bundle\CatalogBundle\AttributeType\AttributeTypes;
 use Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\ProductRepository;
 use Pim\Bundle\CatalogBundle\Model\AssociationTypeInterface;
 use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
@@ -181,7 +182,7 @@ class PublishedProductRepository extends ProductRepository implements PublishedP
     {
         $qb = $this->createQueryBuilder('pp');
 
-        if ($option->getAttribute()->getAttributeType() === 'pim_catalog_simpleselect') {
+        if ($option->getAttribute()->getAttributeType() === AttributeTypes::OPTION_SIMPLE_SELECT) {
             $qb
                 ->innerJoin('pp.values', 'ppv', 'WITH', $qb->expr()->eq('ppv.option', $option->getId()));
         } else {